refactor: rename e2e testing scripts
